Show CONNECTING and SELECT_PHONE_ACCOUNT states as bluetooth-dialing.
When a bluetooth device asks us to make a call, we make a new Call
instance in the CONNECTING state and ask the ConnectionService to
place the call. Once the ConnectionService places the call the state
transitions to DIALING.

The bluetooth device expects there to be a call after it asks us to
place it, but we return zero calls if the device asks us prior to
transitioning to DIALING. This is because we treat CONNECTING as an
IDLE state for bluetooth devices.

The reason we treated CONNECTING as an IDLE state was to accommodate
aggressive end-call behavior from apps which rewrite a phone numbers
when making phone calls.  We have since fixed that problem in
other ways and it is no longer necessary to treat CONNECTING or
SELECT_PHONE_ACCOUNT as IDLE states. This should fix problems we are
seeing where BT devices show zero calls if they query our system during
the CONNECTING state.

Bug: 24093455
Change-Id: I129711f6b9f6b7e32fc0bfa338b1e7fa5965e378