If reusing call, don't check makeRoomForCall
Currently, if a call is reused in startOutgoingCall (in GV, for example)
during the making of a conference call, makeRoomForCall will return false
mistakenly because the call has already transitioned from
NEW -> CONNECTING state.

During the making of the new Call, makeRoomForCall was already called
once to be sure that the new call is valid. We can safely avoid making
this call to makeRoomForCall again and returning an incorrect "false."

Bug: 22980265
Change-Id: Id6fbad299f75a92d6c5b1be06e38611f8113c822