Add timer to make in-call ui reflect bluetooth state
Adds a timeout of 5 seconds to connectBluetooth. Within those 5 seconds,
it will be assumed that audio is being routed through bluetooth. An
additional check is done at the end of the 5 seconds to see if bluetooth
audio is connected. If not, the in-call ui is updated to reflect this
fact.

This differs from the previous behavior in that there was no additional
check done at the end of the 5 seconds previously.

Bug: 25378995
Change-Id: I958765ce3f6facfe93fb0c54e77b0130c7fde408