Refactor CreateConnectionProcessor
Currently, if a ConnectionManager refuses a connection from a Call, this
code will set a flag and call process() again. This finds all of
possible PhoneAccounts again (without the addition of the
ConnectionManager). In order to avoid this redundancy, a queue is utilized to
have the CreateConnectionProcessor fall back to another account if the
connection fails.

This change only changes how a ConnectionManager is being found/used.

Also, tests have been added for CreateConnectionProcessor

Change-Id: Id7479e5337a812fdae83411fb2fbc248d8c1a7da