Introduces sessions into Telecom Logging
Sessions are introduced into the Telecom logging framework in order
to better trace code execution paths in the Telecom logs. Mostly,
this code keeps track of threads that are currently executing code in
Telecom and when a log or event is recorded, the unique session
information is appended to the log/event.

Change-Id: Ied3940fbb33e097636a99571e946dd40a154d64b