Do not route audio to speaker if Bluetooth audio is connected
In some devices a magnetic charging port doubles as a dock connector
and if user is charging phone and the charge port is identified as a
dock the voice audio will be routed through the phone speaker even if
Bluetooth audio is connected.

Priorities Bluetooth over dock and route audio to Bluetooth if
connected.

Bug: 28914501
Author: Mathias Jeppsson <mathias.jeppsson@sonymobile.com>
Change-Id: Ib09fa96f8d78a8a8f202fca69c92ee88cb181f7a