Fix issue where CEP participants cannot be disconnected.
Existing connections are added in Telephony for the CEP participants.
In that connection service, a unique id generated using the package name
of the connection service and a sequential number.  This is passed here
into callid.  HOWEVER, I mistakenly did not use it when making the new
call in Telecom.  Hence Telecom would be referring to a call by one ID
and the connection service by another.

Bug: 25779129
Change-Id: I72b3aea3d49f6852a2c4d7a66e2c22494f651d7a