Fixes Permission Error in TelecomSystemTest
Recent changes in TelecomSystemTest aim to test a new
KEY_INITIATING_USER Extra set in
new UserCallIntentProcessor(Context context,...).processIntent(...).
Previously, context was not getting set to the correct MockContext,
which was then improperly sending out an intent to the system.
This change uses the correct MockContext and verifies the mocked
sendBroadcastAsUser intent was called before continuing tests.

Change-Id: If09a5fe6cf09fe7bfb0ce8b7e4e0890465bc8aea