No longer start visible subsessions in Logging
Currently, if Log.startSession() is called when a session is
already active, a new subsession will be created (along with
the corresponding Log messages in logcat). This can cause confusion
in the Logging messages because a subsession was not started
explicitly and also adds un-needed verbosity.

This change creates "invisible" subsessions instead that still
track the start/end of the invisible subsession, but do not
report themselves in the log messages.

Also, to help with verbosity in the logs, the log messages
for createSubsession() have moved down to verbose.

Change-Id: Ifc1638b0bb197f20b6baa70293df5bc389dd1fc2