A stale session that was occurring has been corrected
When the ConnectionServiceWrapper calls the removeCall method,
an invisible session was not being created, however the Logging system
was still treating one of the subsessions as an invisible session. To
fix this problem, Log.endSession now checks the intended invisible
session flag instead of just checking if the thread ids are the same.

Also, LogTest now does not print messages during unit tests.

Change-Id: I22888f19263d8d0f7c235b2d50aa05ab5d130ae7