Update isWorkCall to work with multi-SIM devices
Updates InCallController.toParcelableCall(...) to only check if a call
is a Work call after the user has selected the Phone Account that will
be used to make the call. Before the user has selected the Phone Account
to use, the call will be set as not a Work call and will be updated to
the correct state after the Phone Account is chosen.

Bug: 26484778
Change-Id: I7162ab5ee2ac117a2ba02bc0dc00a0f71fd59e7b