Treat the call as personal call if no work dialer is installed
For ACTION_DIAL intent, we will fallback to personal side if there is no
work dialer installed.
But for ACTION_CALL intent, it is handled in work profile and we consider
it as a work call although there is no work dialer installed. We would
now like to make it consistent and consider it as a personl call in this
case. It avoids situation that we insert the call in work side but user has
no work dialer to take a look of the call log.

BUG=26634502

Change-Id: I65cd0b552736b436e60dcaf34d0b2cb055b581eb