Allow dialer to show missed call notification
This CL allows the default dialer app to show the missed call
notification. This is important because:
 - the dialer app can show the same caller ID info that
   it would normally show in the call log
 - the dialer app can make sure the UI is consistent with the rest of
   the app

With this change, telecom never modifies a call's NEW or IS_READ
property. Instead, the dialer app is responsible for clearing the NEW
field once the user dismisses the notification. Similarly, the IS_READ
field is set once the user actually views the call in the dialer
calllog.

Note, this behavior is only implemented for the primary / non-work
profiel user. I'll work on implementing this correctly for the
non-primary user in a follow up CL.

BUG: 22857261
Change-Id: Ifae82d3036a278e5619edc8c766a1d0781296e72