Report phone-state when calls are added.
PhoneStateBroadcaster was previously updating the phone state when calls
change their state.  However, it was not sending updates when the calls
were added unless the call was a ringing call.

This change makes it so that any call changes (addCall, removeCall,
stateChange) cause the phone-state to recalculate and send an update.

Bug: 26693196

Change-Id: I04be5d0338e3be9b769c9366e6ac188075dadb19