Fix phone account is disabled everytime after reboot
There is a bug in the code that deserializes phone account from xml,
all phone accounts without extra are disabled after reboot!

The tag next to "extra", which is "enabled", is swallowed in the
readBundle function and hence the value is ignored.

Also update test case to cover this case.

BUG=26792269

Change-Id: I2387c7aed93b79d636ede2c9da4256b3272e86b0