Fix audio routing issue and flaky test
There was a potential issue in CallAudioRouteManager where the
transition from a quiescent state to its corresponding active state
would not be communicated to the in-call service. This has been fixed.

This change also changes InCallController to not broadcast current state
to all bound in-call services when a new service is bound.
VideoCallTests was adjusted to take this into account.

Bug: 26827840
Change-Id: Ibda3e2c783b77d82a7337fa39f4675dac22ae071