Add call should always be present in CDMA calls.
This is reverting Change I2e4b396f5665b5517275766daa722866cf92d68c

Initially, the idea was that CDMA could not place additional calls while
a conference was happening.  This is not always the case.  CDMA
conferences exist in two states: merged and unmerged.  Also, a call can
disconnect from the conference and the conference will continue as a 1:1
call.

In the case where the conference is "merged" or a party has
disconnected, the "add-call" option will function fine. This is why we
are adding the add-call back.

Bug: 22850789

Change-Id: Ib6bafed528cb84d46b46b11313db902d9aba4292