Skip DISALLOW_OUTGOING_CALLS checking in managed profile.
In pre-N, DISALLOW_OUTGOING_CALLS is not checked in managed profile as
ACTION_CALL is always forwarded to personal side.

In N, ACTION_CALL is now handled in managed profile and hence
we do check DISALLOW_OUTGOING_CALLS. It does not align with the javadoc in
http://developer.android.com/intl/ja/reference/android/os/UserManager.html#DISALLOW_OUTGOING_CALLS.

After offline discussion, we decide to keep the pre-N behavior,
because user can always copy and paste the phone number from work side
to personal side anyway.

Bug: 27165111
Change-Id: Id696c0ce1f3bb3fff56addc8646e997a3dfb5a42