Let WiredHeadsetManager recognize USB headsets.
WiredHeadsetManager will now determine whether a headset is plugged in
by querying the list of audio devices and matching it against a list of
device types recognized as wired headsets, rather than calling
AudioManager#isWiredHeadsetOn which ignores USB headsets.

Detecting when the list of connected audio devices is now done through
an AudioManager callback instead of listening to an Intent.

Bug: 25110008
Change-Id: I27ac4b835be8a76529c3564df3f8d3d00540cc84