Change priority of call states in determining audio focus state.
In some cases, if an incoming call is answered over an already-active
call, the active call will go into the hold state before the incoming
call exits the ringing state. This causes the CAMSM to rapidly go in and
immediately out of the RINGING state, which ends up disabling call audio
for the newly answered incoming call. The fix is to disable
transitioning to a ringing state when there are holding calls.

There was a previous change made that prioritized ringing calls over
holding/tone-playing calls. This was so that if the active call were
remotely disconnected while an incoming call was on call-waiting, the
ringtone would start playing. This behavior is preserved by setting the
priority as follows:
holding calls > ringing calls > tone playing

Change-Id: I617e7733b47f244f8d668b9e86166143383a29b6
Fixes: 27804164