Add year, month, day to telecom dumpsys.
The dumpsys for Telecom only shows the time of a call in the heading
for each call.  This bit me in b/27729247, as there appears to have been
a call lingering around for a long time that was not disconnected.

Adding the year, month, day to the heading for each call will make it
easier to spot this type of anomaly quickly.

Bug: 27729247
Change-Id: Ia880864bcc6a91d2a758a7750247be9d0f04b3b4