Incorporate parallel call blocking checks
Modify CallsManager to use the IncomingCallFilter class introduced in a
previous CL.
Modifies AsyncBlockCheckTask and CallScreening to implement CallFilter
and removes the logic in those classes that used to execute the filters
in series.
Added tests for block-check and call screening.
Added logging events to track how long each of the call filters take.

Change-Id: Iecd60463fe9ac6dbbd85d57ef6ae2bfccb94d967
Fix: 26883888