Modify outgoing Call logic to support multiple PhoneAccounts
When a Phone call is active, Telecom will use the same
PhoneAccount for any subsequent outgoing call made during the life of
the active foreground call. This change constructs a list of possible
PhoneAccounts that a user may want, depending on if any of the active
calls are on a SIM PhoneAccount and uses that list to determine if the
chosen PhoneAccount is available.

Change-Id: Ib8a2c90c6c6631524200dda040515f7f5b4e7f0f
Fix: 18836240