Modify CallAudioManager's audio speedup functionality
For certain calls, CallAudioManager will put the audio stream into the
in-call state before the connection service has acknowledged the call as
active. This was previously done by hardcoding a state transition in
CallAudioModeStateMachine. However, this had the potential to leave a
call stuck in CallAudioManager if the connection service subsequently
disconnected the call without setting to active after the user has
requested to answer the call.

This change solves this problem by changing the call state transition
code to go through all of its internal call tracking lists and remove the
call instead of only looking at the one corresponding to the state that
call was in.

Change-Id: Ica4a182fdd210a17ca58402062c8358d193d90c8
Fix: 28114043