Synchronize the bg/fg call state changes.
Chain the background call actions to the completion of the
foreground call's state change instead of running them in a Dialog
parallel to the manager's state machine.

Use the existing wrappers, which already do validity checks, instead
of invoking the operations directly in the Call object.

Also, react to call state changes when the dialog is live. If
one of the calls we're making a decision on goes away, pick up
the other.

TICKET: CYNGNOS-2688

Change-Id: I1197f02587a518b7fed0cd74a7fbca59a02957a4