Add check for "Silent" Uri.EMPTY Ringtone Uri
Currently we do not check for the ringtone Uri Uri.Empty, which is
the "Silent" ringtone. Since we did not check for this case, we would
play a "Dinging" sound instead of nothing when we queried for the
Ringtone associated with the Uri.

This change checks to see if the Ringtone Uri is Uri.EMPTY before
playing it.

Change-Id: Iab51d43184f29caf58f5333a7c893c26cd9bca0a
Fix: 28912183