Add Group Id tracking in PhoneAccountRegistrar
Adds Group Ids to the DefaultPhoneAccountHandle in Telecom in order to
keep track of the group that the currently registered PhoneAccounts
belong to. If registerPhoneAccount is called with a new PhoneAccount
that has the same group Id as a currently registered PhoneAccount, then
remove the registered PhoneAccount and register the new PhoneAccount
with the old PhoneAccount's defaults and enabled status. This allows us
to transparently change PhoneAccounts and keep the calling account
defaults the same.

Bug: 28173788
Change-Id: Ida44e348903d2ba6b8b8f8cc64a5b13cff68c1c2