Fix corruption in InCallTonePlayer
InCallTonePlayer uses a static variable to keep track of the number of
tones playing, which is incremented in startTone and decremented in
cleanUpTonePlayer. If something calls the start() method directly,
sTonesPlaying will not be correctly incremented.

This change removes the instance where start() was called directly and
changes start() on InCallTonePlayer to be a no-op with a warning
message.

Change-Id: I03a20f89fb4458e29f67ba327a31a8f82ff87423
Fixes: 29581234