Allow for Reconnect from Connection Manager
There are currently two situations where Telecom will redial on another
ConnectionService's PhoneAccount: 1) An Emergency Call has been placed,
Telecom connects successfully to a ConnectionService, and the
ConnectionService returns an ERROR message; and 2) A ConnectionManager
is being used to place a call and the ConnectionManager rejects the
Call with a onCreateConnectionFailure callback.

In case 1), we disconnect to the ConnectionService connect to another
emergency call capable PhoneAccount to place the call.
In case 2), we try to place the call directly on the ConnectionService
that the PhoneAccount is registered to.

This change reverts part of a previous change that broke case 2).

Bug: 29877938
Change-Id: I58badca94bb893f8ec797fc16ecb66f8ca9acabd