Fix ComponentName NPE when binding to InCallService
Currently, if the PackageManager can not find the component that is
set to be bound, it returns null. This creates a null InCallServiceInfo,
which creates a NPE when getComponentName is called during bind. We now
never return null and instead try to bind to the ComponentName anyway if
the PackageManager returns null.

Bug: 30150689
Change-Id: I463332b1707a76d8b018aaaacbca542c91948dbf