When bg call does not support hold, auto-unhold on disconnect of fg call.
When disconnecting a call and the new foreground call is a held call
that does not support hold, auto-unhold that call.  This handles the
scenario where a carrier (for example) does not support hold, and a call
disconnect causes a background call to be on hold.  Since the carrier does
not support hold/unhold, the UX will provide no means to unhold the call.

Bug: 28813047
Change-Id: I4fe7410e5ef49297bdfb968d2e03a1594fa0d897