InCallController: don't strip can send SMS capability on MSIM
With MSIM devices, if a default SMS subscription is required with AOSP,
so logic existed to remove that flag if a default SMS sub wasn't
selected.

We support always ask, and we also now disable always ask if the default
SMS app does not support it. We can leverage this logic here and just
pass the flag along and let it be handled in the UI layer.

This also covers the edge case of when Hangouts is selected as the defualt
(does not support always ask), and the user DID have always ask selected,
and then received an incoming call, and selects to respond with an SMS:
we are still prompted to select a new SIM to send the SMS with.

Ticket: CYNGNOS-3225

Change-Id: Id5bf9c60e18278a2acabc893398e27f3ba2d4700
Signed-off-by: Roman Birg <roman@cyngn.com>
(cherry picked from commit 7130db321167e3dc6e0a2bf99a844e6233dd649a)