Auto reject video requests for secondary user.
- In Telecom, when the ConnectionService sends informs us of an upgrade
to video request, if the phone account for the call does not support video
calling, then we will automatically send a response back saying the user
wants to stick with audio-only.
- Also, it turns out the Call object already knew whether it supports video
or not, so I was able to clean up the change I made previously to prevent
video calling on the secondary user.

Bug: 30600205
Change-Id: I64a29933d9da3f443dbad5fbd058f4b5735cfc51