Prevent notification vibration when the user is on the earpiece.
When the earpiece route is activated, if the user has not already enabled
a notification interruption filter, we will set it to alarms-only.  This
ensures that notifications don't cause the phone to vibrate while in
earpiece mode.

If the user changes the interruption filter during the call we will keep
their new chosen filter when the earpiece route is disabled.

This notification filtering only takes place for the earpiece route.

Bug: 29962221
Change-Id: Ia332a988f10427a2a0f8315af9f31d854836db9f